+++ galaxy/src/galaxy/graphics/Texture.hpp
@@ -92,7 +92,7 @@ namespace galaxy
 			///
 			/// \param level 2, 4, 8, etc...
 			///
-			void set_anisotropy(const unsigned int level) noexcept;
+			void set_anisotropy(const float level) noexcept;
 
 			///
 			/// Set filter when texture is downscaled in OpenGL.
